requirements.txt
pytest
python-telegram-bot==22.3
notion-client==2.4.0
python-dotenv==1.1.1
email-validator==2.2.0
